requirements.txt
python-telegram-bot==22.3
python-dotenv==1.1.1
peewee==3.18.2
together==1.5.21
